Some tweaks to improve handling *really* large single sequences (> 1Gb). It's still not robust in that situation and not to be recommended.
git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2482 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/gap5/fasta.c
@@ -23,9 +23,9 @@ typedef struct {
     char *seq;
     char *qual;
     int  max_name_len;
-    int  max_seq_len;
-    int  max_qual_len;
-    int  seq_len;
+    size_t  max_seq_len;
+    size_t  max_qual_len;
+    size_t  seq_len;
 } fastq_entry_t;
